requirements.txt
torch>=1.12.0
torchvision
diffusers
accelerate
gradio
tqdm
Pillow